refactor: expand astra-skills with executor, improvement, verify, dynamic_skills, database provider
Extract 6 modules (~2,860 LOC) from runtime/skills/ to the astra-skills crate:

- executor/inline.rs (448 LOC): Inline skill instruction injection
- executor/isolated.rs (556 LOC): Fork/sub-agent skill execution
- improvement.rs (482 LOC): Skill auto-improvement detection
- verify.rs (222 LOC): Skill verification via VerificationRunner
- providers/database.rs (403 LOC): Database-backed SkillProvider adapter
- providers/dynamic_skills/ (748 LOC): 8 bundled dynamic skill templates

Also moves has_inline_shell() utility to astra-skills.

Runtime files become 1-line re-export stubs for backward compatibility.

Test results:
- astra-skills: 310 tests pass
- astra-runtime: 2642 pass, 6 pre-existing failures (network listener tests)
