+++ rust/crates/astra-skills/src/lib.rs
@@ -45,3 +45,31 @@ pub use manifest::{
     EffortLevel, ExecutionContext, LoadedSkill, SkillErrorKind, SkillManifest, SkillSourceKind,
 };
 pub use traits::{SkillError, SkillExecutor, SkillProvider, SkillResolver};
+
+// Runtime skill execution and management
+pub mod executor;
+pub mod improvement;
+pub mod verify;
+
+// Re-export key types
+pub use executor::{
+    InlineSkillExecutor, IsolatedSkillExecutor, SkillExecutionRouter, SkillSubRunExecutor,
+    SubRunResult,
+};
+pub use improvement::{ImprovementProposal, ImprovementTracker, SkillImprovement, TURN_BATCH_SIZE};
+pub use verify::SkillVerifier;
+
+/// Detect inline shell command lines in skill instructions.
+pub fn has_inline_shell(instructions: &str) -> bool {
+    instructions.lines().any(|line| {
+        let t = line.trim_start();
+        if t == "!" {
+            return true;
+        }
+        if let Some(rest) = t.strip_prefix('!') {
+            rest.starts_with(char::is_whitespace)
+        } else {
+            false
+        }
+    })
+}
