+++ rust/crates/runtime/src/skills/executor/mod.rs
@@ -1,9 +1,1 @@
-//! Skill execution modules.
-
-pub mod inline;
-pub mod isolated;
-
-pub use inline::InlineSkillExecutor;
-pub use isolated::{
-    IsolatedSkillExecutor, SkillExecutionRouter, SkillSubRunExecutor, SubRunResult,
-};
+pub use astra_skills::executor::*;
