+++ rust/crates/runtime/src/skills/providers/mod.rs
@@ -1,18 +1,8 @@
-//! Skill providers — concrete implementations of the `SkillProvider` trait.
+//! Skill providers — re-export layer.
 //!
-//! Core providers (LocalSkillProvider, BundledSkillProvider, McpSkillProvider) are
-//! now provided by the `astra-skills` crate. This module provides runtime-specific
-//! providers that depend on astra-services:
-//!
-//! | Provider | Source | Priority |
-//! |----------|--------|----------|
-//! | [`DatabaseSkillProvider`](database::DatabaseSkillProvider) | Database (via `SkillService`) | 2 |
-//!
-//! Dynamic skills (bundled skills generated at build time) are registered here
-//! into the BundledSkillProvider from astra-skills.
-
-pub mod dynamic_skills;
+//! Core providers and database provider are now in the `astra-skills` crate.
 
 pub mod database;
+pub mod dynamic_skills;
 
 pub use database::DatabaseSkillProvider;
